% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BDA.R
\name{BDA}
\alias{BDA}
\title{Performs Binary Differential analysis on a seurat object}
\usage{
BDA(
  object,
  contrast.by,
  ident.1,
  ident.2,
  group.by = NULL,
  cell.populations = NULL,
  min.pct = 0.1,
  covariates = NULL,
  p.adjust.method,
  n.cores = 1,
  method = "LR"
)
}
\arguments{
\item{object}{A seurat object}

\item{contrast.by}{Column in which the contrasting cell populations are
defined.}

\item{ident.1}{First cell population of contrasting cell populations.}

\item{ident.2}{Second cell population of contrasting cell populations.}

\item{group.by}{Column in which groups are defined. For each
group Binary Differential analysis will be performed between the contrasting
populations.}

\item{cell.populations}{Groups of cells for which the differential dropout
analysis will be performed separately.}

\item{min.pct}{Only tested genes that were expressed in at least min.pct% of
the cells in either of the respective groups of interest.}

\item{covariates}{Covariates}

\item{p.adjust.method}{Multiple testing correction method.}

\item{n.cores}{Number of cores. When multicore is enabled, progress bar
cannot be displayed.}

\item{method}{which BDA test to use; Logistic regression (LR),
 chi squared test (chisq) Fisher exact (fisher) or  Pearson's Correlation
(phi). Of Note only LR allows to adjust for covariates.}
}
\value{
List of differential dropout analysis results.
}
\description{
Performs Binary Differential analysis on a seurat object
}
\examples{
\dontrun{
library(BDA)
results <- BDA(object = data,
contrast.by = "Status", ident.1 = "COVID", ident.2 = "Healthy",
group.by = "cell.type.coarse", cell.populations = c("RBC","B"),
p.adjust.method = "fdr")}
}
