% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareJaccard.R
\name{compareJaccard}
\alias{compareJaccard}
\title{Calculates the jaccard index between BDA and DEA based on significantly
detected genes}
\usage{
compareJaccard(BDA, DEA, signThreshold, p.adjust.method)
}
\arguments{
\item{BDA}{Results of differential dropout analysis performed with BDA}

\item{DEA}{Results of differential expression analysis performed with the
FindMarkers function of Seurat}

\item{signThreshold}{Threshold at which genes are considered significant}

\item{p.adjust.method}{Multiple testing corrextion method see `p.adjust()`}
}
\value{
jaccard index
}
\description{
Calculates the jaccard index between BDA and DEA based on significantly
detected genes
}
