% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataAD.R
\docType{data}
\name{dataAD}
\alias{dataAD}
\title{Seurat object of AD dataset (Grubman, A. et al 2019)}
\format{
A seurat object
}
\source{
(http://adsn.ddnetbio.com/)
}
\usage{
dataAD
}
\description{
A single-cell atlas of entorhinal cortex from individuals with Alzheimer’s disease reveals cell-type-specific gene expression regulation
}
\keyword{datasets}
