\name{BreastCancer}
\docType{data}
\alias{normal}
\alias{primary}
\alias{meta}
\alias{LCL}

\usage{data(BreastCancer)}
\title{Breast Cancer Data}
\description{
 Cleaned breast cancer data from TCGA. 
 Three datasets: primary, normal and meta.
}
\value{None}
\references{
TCGA(2012)
Comprehensive molecular portraits of human breast tumours.
Nature,490(7418),61-70.
}

\keyword{datasets}
