\name{FSD}
\docType{data}
\alias{FSD}
\alias{Employment2}
\alias{Firm2}
\alias{Job2}
\alias{FirmAge}
\alias{FirmDeathAge}
\alias{FirmDeathSize}
\alias{FirmEmploymentAge}
\alias{FirmEmploymentSize}
\alias{FirmJobAge}
\alias{FirmJobSize}
\alias{FirmSize}

\title{Firm size data}
\description{
 2014 US Private-sector firm size data.
}

\references{
 https://www.sba.gov/advocacy/firm-size-data, Accessed on 2018-11-16
}

\keyword{datasets}
