\name{Pain}
\docType{data}
\alias{Pain}
\alias{pain}

\title{Pain data}
\description{
 Pain data using VAS.
}
\format{A data frame with 203 records on 17 variables.
  \tabular{lll}{
    \code{group}  \tab character  \tab 'control' or 'treatment' group \cr
    \code{t01-t07}   \tab numeric  \tab VAS measures at T0 from diary \cr
    \code{t11-t17}   \tab numeric  \tab VAS measures at T1 from diary \cr
    \code{rvas0}   \tab numeric  \tab recalled VAS average at T0 \cr
    \code{rvas1}   \tab numeric  \tab recalled VAS average at T1
  }
}

\references{
 To be updated
}

\keyword{datasets}
