% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateCorrelation.R
\name{estimateCorrelation}
\alias{estimateCorrelation}
\title{Calculates the correlation coefficient between logOR and logFC of a
BDA and DEA respectively}
\usage{
estimateCorrelation(BDA, DEA, method = "pearson")

estimateCorrelation(BDA, DEA, method = "pearson")
}
\arguments{
\item{BDA}{Results of differential dropout analysis performed with BDA}

\item{DEA}{Results of differential expression analysis performed with the
FindMarkers function of Seurat}

\item{method}{a character string indicating which correlation coefficient
is to be computed. See `cor()`}
}
\value{
jaccard index

jaccard index
}
\description{
Calculates the correlation coefficient between logOR and logFC of a
BDA and DEA respectively

Calculates the correlation coefficient between logOR and logFC of a
BDA and DEA respectively
}
