% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimatePlot.R
\name{estimatePlot}
\alias{estimatePlot}
\title{Generates a ggplot of the logOR versus the logFC of a BDA and DEA respectively}
\usage{
estimatePlot(BDA, DEA)
}
\arguments{
\item{BDA}{Results of differential dropout analysis performed with BDA}

\item{DEA}{Results of differential expression analysis performed with the
FindMarkers function of Seurat}
}
\value{
ggplot
}
\description{
Generates a ggplot of the logOR versus the logFC of a BDA and DEA respectively
}
