\name{EUFirmSize}
\docType{data}
\alias{EUFirmSize}

\usage{data(EUFirmSize)}
\title{Firm size data of 10 EU countries}
\description{
Mining and quarrying firm employment sizes.
Classes: "0-9","10-19","20-49","50-249","250+".
}
\value{
\item{Class}{Firm size classes}
\item{Year20xx}{Firm size data of 10 EU countries for year 20xx.}
}
\references{
Eurostat, Enterprises in Europe, Data 1994-95, fifth report Edition, European
  Commission, Brussels, 1998.
}

\examples{
 data(EUFirmSize)
 head(EUFirmSize)
}

\keyword{datasets}
