% This is a template R document file
% Distributed under GPL 3 or later

\name{fitGLD2bdata}
\alias{fitGLD2bdata}

\title{Fit FMKL GLD to Binned Data}
\description{
   Fit FMKL GLD to binned data.
   }
\usage{
  fitGLD2bdata(x, weights)
}
\arguments{
  \item{x}{histogram-type data such as histogram or bdata.}
  \item{weights}{to weight the squared distance for different classes.}
}

\author{
  B. Wang  \email{bwang831@gmail.edu}
}


\references{
  AS 254, ...  
}
\examples{
 # To be updated.
 x = rexp(100,1)
}
\keyword{smooth}

