% This is a template R document file
% Distributed under GPL 3 or later

\name{fit.GPD}
\alias{fit.GPD}

\title{Fit a Generalized Pareto Distribution to Binned Data}

\description{
   Fit a Generalized Pareto distribution to binned data.
   }
\usage{
  fit.GPD(x)
}
\arguments{
  \item{x}{grouped data}
}

\author{
  B. Wang  \email{bwang@southalabama.edu}
}


\references{
  AS 254, ...  
}
\examples{
 # To be updated.
 x = rexp(100,1)
}
\keyword{smooth}

