% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LR.R
\name{LR}
\alias{LR}
\title{LR performs BDA with a logistic regression}
\arguments{
\item{dataLR}{Temporary binary matrix}

\item{covariates}{Temporary samplesheet}
}
\description{
LR performs BDA with a logistic regression
}
\keyword{internal}
