% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcBDA.R
\name{mcBDA}
\alias{mcBDA}
\title{mcBDA multi core BDA}
\arguments{
\item{tmp_binaryMatrix}{Temporary binary matrix}

\item{tmp_samplesheet}{Temporary samplesheet}

\item{contrast.by}{Contrast column}

\item{ident.1}{First cell population of contrasting cell populations.}

\item{ident.2}{Second cell population of contrasting cell populations.}

\item{covariates}{covariates}

\item{cores}{Number of cores}
}
\description{
mcBDA multi core BDA
}
\keyword{internal}
