% This is a template R document file
% Distributed under GPL 3 or later

\name{perm.test.NGS}
\alias{perm.test.NGS}

\title{A permutation test for DEG detection}

\description{
  To perform a permutation test to check whether a gene is differentially expressed.
}

\usage{
  perm.test.NGS(x,y, alternative = "two.sided", iter = 1001)
}
\arguments{
  \item{x,y}{The expressions of a gene in two populations}
  \item{alternative}{Type of test}
  \item{iter}{Iterations for the test.}

}

\author{
  B. Wang  \email{bwang@southalabama.edu}
}

\examples{
 # To be updated.
}
\keyword{smooth}

