% This is a template R document file
% Distributed under GPL 3 or later

\name{q.test}
\alias{q.test}
\alias{q.test.default}
\alias{q.test.NGS.lognormal}

\title{Q-test}

\description{
  To perform a test on a quantile of a distribution.
}

\usage{
  q.test(x,p,q)
}
\arguments{
  \item{x}{An R object.}
  \item{p,q}{The percentage and quantile.}
}

\author{
  B. Wang  \email{bwang@southalabama.edu}
}

\examples{
 # To be updated.
}
\keyword{test}

