\name{LCL}
\docType{data}
\alias{LCL}

\usage{data(LCL)}
\title{Two selected profiles}
\description{
 This dataset contains two selected gene expression profiles from Gene Expression Omnibus (GEO) by NIH NCBI (accession number ``GSE19480'').
}

\references{
Pickrell JK, Marioni JC, Pai AA, Degner JF, Engelhardt BE, Nkadori E, Veyrieras JB, Stephens M, Gilad Y, Pritchard JK.
 Understanding mechanisms underlying human gene expression variation with RNA sequencing.
 Nature. 2010 Apr 1;464(7289):768-72. 
}

\keyword{datasets}
