% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scBDA.R
\name{scBDA}
\alias{scBDA}
\title{scBDA single core BDA}
\arguments{
\item{tmp_binaryMatrix}{Temporary binary matrix}

\item{tmp_samplesheet}{Temporary samplesheet}

\item{contrast.by}{Contrast column}

\item{ident.1}{First cell population of contrasting cell populations.}

\item{ident.2}{Second cell population of contrasting cell populations.}

\item{covariates}{covariates}
}
\description{
scBDA single core BDA
}
\keyword{internal}
