\name{bda}
\alias{bda}
\encoding{UTF-8}
\title{
 Binned (Grouped) Data Analysis
}
\description{
This package collects algorithms for binned (grouped) data analysis.
}

\references{
Wang B. (2020).
\emph{A Zipf-plot based normalization method for high-throughput RNA-seq
  data}. 
PLOS ONE 15(4): e0230594.

Wang, B. and Wang, X-F. (2015).
\emph{Fitting The Generalized Lambda Distribution To Pre-binned Data}.
Journal of Statistical Computation and Simulation. Vol. 86 , Iss. 9, 1785-1797.

Ge C, Zhang S-G, Wang B (2020) Modeling the joint distribution of firm size and firm age based on grouped data. PLoS ONE 15(7): e0235282. doi.org/10.1371/journal.pone.0235282

Wang, B. and Wertelecki, W. (2013) Density Estimation for Data With Rounding Errors, Computational Statistics and Data Analysis, 65: 4-12. doi: 10.1016/j.csda.2012.02.016.

Wang, X-F. and Wang, B. (2011) Deconvolution Estimation in Measurement Error Models: The R Package decon, Journal of Statistical Software, 39(10), 1-24.

}
\keyword{stat}
