\name{bda}
\docType{package}

\alias{bda-package}
\alias{print.bdata}

\title{Binned Data Analysis}

\description{

  This package collects algorithms for density estimation based on
pre-binned data, left-/right- or interval censored data, and data
contaminated by measurement errors.

}

\author{
  B. Wang  \email{bwang@southalabama.edu}
}

\keyword{stats}

